backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
groq>=0.4.0
pytesseract>=0.3.0
Pillow>=9.0.0
PyMuPDF>=1.20.0
pdfminer.six>=20220524
python-dotenv>=1.0.0
requests>=2.28.0